+++ packages/opentelemetry-node/CHANGELOG.md
@@ -8,6 +8,7 @@
   of instrumentations. See <https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/plugins/node/opentelemetry-instrumentation-mysql2#readme>
 - feat: Add `@opentelemetry/instrumentation-cassandra-driver` to the default set
   of instrumentations. See <https://github.com/open-telemetry/opentelemetry-js-contrib/blob/main/plugins/node/opentelemetry-instrumentation-cassandra#readme>
+- chore: Test that the native instrumentation in `@elastic/elasticsearch@8.15.0` and later works.
 
 
 ## v0.6.0